refactor(shikomi-infra): split integration_test.rs into files under 500 lines
ペガサス指摘（PR #17 再レビュー）: integration_test.rs 988行 → 500行制限違反。

tests/helpers/mod.rs に共有ヘルパー（make_repo, make_plaintext_vault 等）を集約。
tests/integration_roundtrip.rs: TC-I01, I02, I08, I09, I16〜I19 (246行)
tests/integration_error.rs: TC-I03〜I05, I13〜I15 (315行)
tests/integration_system.rs: TC-I07, I10〜I12, I20〜I23 (380行)
全ファイル500行以内厳守。integration_windows.rs (TC-I24〜I27) は変更なし。
